TaskPulse/requirements.txt
amqp==5.3.1
asgiref==3.10.0
astroid==3.3.11
billiard==4.2.3
black==25.11.0
celery==5.5.3
certifi==2025.11.12
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.11.0
cron_descriptor==2.0.6
dill==0.4.0
distlib==0.4.0
Django==5.2.8
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-filter==25.2
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.0
flake8==7.3.0
flake8-isort==7.0.0
identify==2.6.15
idna==3.11
iniconfig==2.3.0
isort==6.1.0
kombu==5.5.4
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.4.0
prompt_toolkit==3.0.52
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.9
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
PyYAML==6.0.3
redis==5.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
virtualenv==20.35.4
wcwidth==0.2.14
